requirements.txt
sentence-transformers
faiss-cpu
numpy
groq
openai
dotenv